refactor: add email to otp header
